requirements.txt
asgiref==3.7.2
asttokens==2.4.1
colorama==0.4.6
coverage==7.4.1
decorator==5.1.1
Django==5.0.1
django-email-users==1.1.0
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.0
executing==2.0.1
flake8==7.0.0
ipython==8.20.0
jedi==0.19.1
matplotlib-inline==0.1.6
mccabe==0.7.0
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.11.1
pyflakes==3.2.0
Pygments==2.17.2
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.14.1
typing_extensions==4.9.0
tzdata==2023.4
wcwidth==0.2.13
